Refactor view models into partial classes
